refactor VideoQualityPatch
