+++ app/src/main/java/app/revanced/integrations/patches/HideCreateButtonPatch.java
@@ -10,12 +10,9 @@ public class HideCreateButtonPatch {
 
     //Used by app.revanced.patches.youtube.layout.createbutton.patch.CreateButtonRemoverPatch
     public static void hideCreateButton(View view) {
-        String message = SettingsEnum.CREATE_BUTTON_SHOWN.getBoolean() ? "Create button: Shown" : "Create button: Hidden";
+        boolean show = SettingsEnum.CREATE_BUTTON_SHOWN.getBoolean();
+        String message = show ? "Create button: Shown" : "Create button: Hidden";
         LogHelper.debug(HideCreateButtonPatch.class, message);
-        if (SettingsEnum.CREATE_BUTTON_SHOWN.getBoolean()) {
-            view.setVisibility(View.VISIBLE);
-        } else {
-            view.setVisibility(View.GONE);
-        }
+        view.setVisibility(show ? View.VISIBLE : View.GONE);
     }
 }
