+++ app/src/main/java/app/revanced/integrations/patches/HideInfoCardSuggestionsPatch.java
@@ -1,10 +1,12 @@
 package app.revanced.integrations.patches;
 
+import android.view.View;
+
 import app.revanced.integrations.settings.SettingsEnum;
 
 public class HideInfoCardSuggestionsPatch {
 
     public static int hideInfoCardSuggestions() {
-        return SettingsEnum.INFO_CARDS_SHOWN.getBoolean() ? 0 : 8;
+        return SettingsEnum.INFO_CARDS_SHOWN.getBoolean() ? View.VISIBLE : View.GONE;
     }
 }
