+++ app/src/main/java/app/revanced/integrations/patches/HideShortsButtonPatch.java
@@ -9,14 +9,14 @@ import app.revanced.integrations.utils.LogHelper;
 
 public class HideShortsButtonPatch {
 
-    //Todo: Switch BooleanPreferences to Settings class
     //Used by app.revanced.patches.youtube.layout.shorts.button.patch.ShortsButtonRemoverPatch
     public static void hideShortsButton(View view) {
         if (lastPivotTab != null && lastPivotTab.name() == "TAB_SHORTS") {
-            String message = SettingsEnum.SHORTS_BUTTON_SHOWN.getBoolean() ? "Shorts button: shown" : "Shorts button: hidden";
+            boolean show = SettingsEnum.SHORTS_BUTTON_SHOWN.getBoolean();
+            String message = show ? "Shorts button: shown" : "Shorts button: hidden";
             LogHelper.debug(HideShortsButtonPatch.class, message);
-            if (!SettingsEnum.SHORTS_BUTTON_SHOWN.getBoolean()) {
-                view.setVisibility(View.GONE);
+            if (!show) {
+                view.setVisibility(show ? View.VISIBLE : View.GONE);
             }
         }
     }
