+++ app/src/main/java/app/revanced/integrations/patches/NewActionbarPatch.java
@@ -4,8 +4,7 @@ import app.revanced.integrations.settings.SettingsEnum;
 
 public class NewActionbarPatch {
 
-    //ToDo: Write Patch for it.
-    //See https://drive.google.com/file/d/1Jg2WK9wwSABCiIcqclzhedy3J3RCf3Hn/view?usp=sharing for where it needs to be used.
+    //Used by app.revanced.patches.youtube.layout.widesearchbar.patch.WideSearchbarPatch
     public static boolean getNewActionBar() {
         return SettingsEnum.USE_NEW_ACTIONBAR.getBoolean();
     }
