+++ app/src/main/java/app/revanced/integrations/patches/VideoSpeedPatch.java
@@ -1,9 +1,4 @@
-package app.revanced.integrations.videoplayer.videosettings;
-
-
-import app.revanced.integrations.settings.SettingsEnum;
-import app.revanced.integrations.utils.LogHelper;
-import app.revanced.integrations.utils.ReVancedUtils;
+package app.revanced.integrations.patches;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
@@ -11,8 +6,12 @@ import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Iterator;
 
-/* loaded from: classes6.dex */
-public class VideoSpeed {
+import app.revanced.integrations.settings.SettingsEnum;
+import app.revanced.integrations.utils.LogHelper;
+import app.revanced.integrations.utils.ReVancedUtils;
+
+public class VideoSpeedPatch {
+
     public static final float[] videoSpeeds = {0.25f, 0.5f, 0.75f, 1.0f, 1.25f, 1.5f, 1.75f, 2.0f, 3.0f, 4.0f, 5.0f};
     private static Boolean userChangedSpeed = false;
 
@@ -23,9 +22,9 @@ public class VideoSpeed {
             return speed;
         }
         ReVancedUtils.setNewVideo(false);
-        LogHelper.debug(VideoSpeed.class, "Speed: " + speed);
+        LogHelper.debug(VideoSpeedPatch.class, "Speed: " + speed);
         float preferredSpeed = SettingsEnum.PREFERRED_VIDEO_SPEED.getFloat();
-        LogHelper.debug(VideoSpeed.class, "Preferred speed: " + preferredSpeed);
+        LogHelper.debug(VideoSpeedPatch.class, "Preferred speed: " + preferredSpeed);
         if (preferredSpeed == -2.0f) {
             return speed;
         }
@@ -49,18 +48,18 @@ public class VideoSpeed {
         int index = 0;
         while (it.hasNext()) {
             float streamSpeed2 = it.next();
-            LogHelper.debug(VideoSpeed.class, "Speed at index " + index + ": " + streamSpeed2);
+            LogHelper.debug(VideoSpeedPatch.class, "Speed at index " + index + ": " + streamSpeed2);
             index++;
         }
         int speed3 = -1;
         for (float streamSpeed3 : iStreamSpeeds) {
             if (streamSpeed3 <= preferredSpeed) {
                 speed3++;
-                LogHelper.debug(VideoSpeed.class, "Speed loop at index " + speed3 + ": " + streamSpeed3);
+                LogHelper.debug(VideoSpeedPatch.class, "Speed loop at index " + speed3 + ": " + streamSpeed3);
             }
         }
         if (speed3 == -1) {
-            LogHelper.debug(VideoSpeed.class, "Speed was not found");
+            LogHelper.debug(VideoSpeedPatch.class, "Speed was not found");
             speed2 = 3;
         } else {
             speed2 = speed3;
@@ -69,14 +68,14 @@ public class VideoSpeed {
             Method[] declaredMethods = qInterface.getClass().getDeclaredMethods();
             for (Method method : declaredMethods) {
                 if (method.getName().length() <= 2) {
-                    LogHelper.debug(VideoSpeed.class, "Method name: " + method.getName());
+                    LogHelper.debug(VideoSpeedPatch.class, "Method name: " + method.getName());
                     try {
                         try {
                             method.invoke(qInterface, videoSpeeds[speed2]);
                         } catch (IllegalAccessException | IllegalArgumentException | InvocationTargetException ignored) {
                         } catch (Exception e6) {
                             e = e6;
-                            LogHelper.printException(VideoSpeed.class, e.getMessage());
+                            LogHelper.printException(VideoSpeedPatch.class, e.getMessage());
                             return speed2;
                         }
                     } catch (Exception ignored) {
@@ -86,7 +85,7 @@ public class VideoSpeed {
         } catch (Exception e10) {
             e = e10;
         }
-        LogHelper.debug(VideoSpeed.class, "Speed changed to: " + speed2);
+        LogHelper.debug(VideoSpeedPatch.class, "Speed changed to: " + speed2);
         return speed2;
     }
 
@@ -98,15 +97,15 @@ public class VideoSpeed {
         int i = 0;
         if (!ReVancedUtils.isNewVideoStarted() || userChangedSpeed) {
             if (SettingsEnum.DEBUG.getBoolean() && userChangedSpeed) {
-                LogHelper.debug(VideoSpeed.class, "Skipping speed change because user changed it: " + speed);
+                LogHelper.debug(VideoSpeedPatch.class, "Skipping speed change because user changed it: " + speed);
             }
             userChangedSpeed = false;
             return -1.0f;
         }
         ReVancedUtils.setNewVideo(false);
-        LogHelper.debug(VideoSpeed.class, "Speed: " + speed);
+        LogHelper.debug(VideoSpeedPatch.class, "Speed: " + speed);
         float preferredSpeed = SettingsEnum.PREFERRED_VIDEO_SPEED.getFloat();
-        LogHelper.debug(VideoSpeed.class, "Preferred speed: " + preferredSpeed);
+        LogHelper.debug(VideoSpeedPatch.class, "Preferred speed: " + preferredSpeed);
         if (preferredSpeed == -2.0f) {
             return -1.0f;
         }
@@ -137,7 +136,7 @@ public class VideoSpeed {
         int index = 0;
         for (Float iStreamSpeed : iStreamSpeeds) {
             float streamSpeed2 = iStreamSpeed;
-            LogHelper.debug(VideoSpeed.class, "Speed at index " + index + ": " + streamSpeed2);
+            LogHelper.debug(VideoSpeedPatch.class, "Speed at index " + index + ": " + streamSpeed2);
             index++;
         }
         int newSpeedIndex = -1;
@@ -145,18 +144,18 @@ public class VideoSpeed {
             float streamSpeed3 = iStreamSpeed;
             if (streamSpeed3 <= preferredSpeed) {
                 newSpeedIndex++;
-                LogHelper.debug(VideoSpeed.class, "Speed loop at index " + newSpeedIndex + ": " + streamSpeed3);
+                LogHelper.debug(VideoSpeedPatch.class, "Speed loop at index " + newSpeedIndex + ": " + streamSpeed3);
             }
         }
         if (newSpeedIndex == -1) {
-            LogHelper.debug(VideoSpeed.class, "Speed was not found");
+            LogHelper.debug(VideoSpeedPatch.class, "Speed was not found");
             newSpeedIndex = 3;
         }
         if (newSpeedIndex == speed) {
-            LogHelper.debug(VideoSpeed.class, "Trying to set speed to what it already is, skipping...: " + newSpeedIndex);
+            LogHelper.debug(VideoSpeedPatch.class, "Trying to set speed to what it already is, skipping...: " + newSpeedIndex);
             return -1.0f;
         }
-        LogHelper.debug(VideoSpeed.class, "Speed changed to: " + newSpeedIndex);
+        LogHelper.debug(VideoSpeedPatch.class, "Speed changed to: " + newSpeedIndex);
         return getSpeedByIndex(newSpeedIndex);
     }
 
@@ -170,5 +169,4 @@ public class VideoSpeed {
             return 1.0f;
         }
     }
-
 }
