+++ app/src/main/java/app/revanced/integrations/sponsorblock/PlayerController.java
@@ -24,11 +24,10 @@ import java.util.TimerTask;
 import app.revanced.integrations.settings.SettingsEnum;
 import app.revanced.integrations.utils.LogHelper;
 import app.revanced.integrations.sponsorblock.player.VideoInformation;
-import app.revanced.integrations.adremover.whitelist.Whitelist;
+import app.revanced.integrations.whitelist.Whitelist;
 import app.revanced.integrations.sponsorblock.objects.SponsorSegment;
 import app.revanced.integrations.sponsorblock.requests.SBRequester;
 import app.revanced.integrations.utils.ReVancedUtils;
-import app.revanced.integrations.utils.SharedPrefHelper;
 
 @SuppressLint({"LongLogTag"})
 public class PlayerController {
